arith is fn('a: Number, 'a) -> 'a — homogeneous, like comparison
Eric's ruling (2026-07-12): mixed arithmetic was a mistake. Both
operands are ONE numeric type and the result IS that type — the
probe-both/commit-widening unification the comparison ops got, now
with RigidCheck riding both phases so a DECLARED 'a: Number formal
never binds from body arithmetic during its def gate (`x + i64:1`
against arbitrary rigid 'a is ill-typed AT THE DEF — the param_knot
promotion semantics are deliberately reversed, and the promotion
machinery is DELETED: numeric_absorbers, promo_eligible, the absorber
obligations, the operand pre-bind, the union result types).

Concrete literals are inference ANCHORS: `2` is always i64, `x * 2.0`
is required when x is f64, and unbound×unbound operands alias into
one cell so `|a, b| a + b` is fn('a, 'a) -> 'a by construction —
polymorphic literals were considered and rejected (each would be a
deferred cell in a system whose whole jul12 arc fought for EARLY
settlement).

Datetime/duration arithmetic is evicted to explicit sys::time
functions carrying the operators' exact semantics: add/sub (datetime ±
duration, saturating at the range limits), add_dur (catchable overflow),
sub_dur (saturating at zero — durations are unsigned), scale (duration
× f64, catchable on negative/NaN/overflow; divide by scaling with 1/x).

Fallout fixed along the way, each a real bug the ruling predicted:
- Builtin-lambda sig matching never checked a BOUND impl tvar's
  binding, and the satisfaction check read only single-conjunct cells
  (the display listing rule) — a sandboxed dynamic-module impl
  inferred at f64 slipped under a fn(i64) -> i64 sig and leaked
  f64:2. through it (dynamic_module1). Bound impl tvars now recurse
  structurally; satisfaction validates EVERY cell conjunct.
- Pack's constraint slot now carries the FULL cell pairs (anonymous
  inference conjuncts included), not the declared-quantifier view —
  the wire transports facts, headers are display.
- The strict arg-boundary check gates on CLOSED formals: an unbound
  formal is the pair's first unification, which tc0 already does with
  proper ordering (re-running it hit the Set-walk's greedy
  first-member bind — radio.gx, which turns out to be broken by an
  OLDER regression, tracked separately).
- first_class_lambdas' generic-param body (`f(y) + 1`) is now
  correctly a compile error; the monomorphic reshape FUSES (annotation
  upgraded to Jit).

This closes the last living witness family: the jul12h mixed-acc
class compile-rejects in BOTH modes, consistent with its direct forms
(pinned: findings/homogeneous-arith-jul2026). m6/b4/f3 all AGREE.

Gates: workspace 2289/0, regress 144/0, detcheck 344/0 flaps, book
examples sweep clean except the pre-existing radio.gx (task #47).

+++ stdlib/graphix-tests/src/lang/select.rs
@@ -45,10 +45,10 @@ run!(looping_select, LOOPING_SELECT, |v: Result<&Value>| match v {
 
 const SELECTSTRUCT: &str = r#"
 {
-  type T = { foo: string, bar: i64, baz: f64 };
-  let x = { foo: "bar", bar: 42, baz: 84.0 };
+  type T = { foo: string, bar: f64, baz: f64 };
+  let x = { foo: "bar", bar: 42.0, baz: 84.0 };
   select x {
-    T as { foo: "foo", bar: 8, baz } => baz,
+    T as { foo: "foo", bar: 8.0, baz } => baz,
     T as { bar, baz, .. } => bar + baz
   }
 }
@@ -89,8 +89,8 @@ run!(match_exhaust1, MATCH_EXHAUST1, |v: Result<&Value>| match v {
 
 const NESTEDMATCH0: &str = r#"
 {
-  type T = { foo: (string, i64, f64), bar: i64, baz: f64 };
-  let x = { foo: ("bar", 42, 5.0), bar: 42, baz: 84.0 };
+  type T = { foo: (string, f64, f64), bar: i64, baz: f64 };
+  let x = { foo: ("bar", 42.0, 5.0), bar: 42, baz: 84.0 };
   let { foo: (_, x, y), .. }: T = x;
   x + y
 }
@@ -104,8 +104,8 @@ run!(nestedmatch0, NESTEDMATCH0, |v: Result<&Value>| match v {
 
 const NESTEDMATCH1: &str = r#"
 {
-  type T = { foo: {x: string, y: i64, z: f64}, bar: i64, baz: f64 };
-  let x = { foo: { x: "bar", y: 42, z: 5.0 }, bar: 42, baz: 84.0 };
+  type T = { foo: {x: string, y: f64, z: f64}, bar: i64, baz: f64 };
+  let x = { foo: { x: "bar", y: 42.0, z: 5.0 }, bar: 42, baz: 84.0 };
   select x {
     T as { foo: { y, z, .. }, .. } => y + z
   }
